make RxJava helpers optional in Relation
