+++ library/src/main/java/com/github/gfx/android/orma/widget/OrmaRecyclerViewAdapter.java
@@ -19,6 +19,7 @@ package com.github.gfx.android.orma.widget;
 import com.github.gfx.android.orma.Relation;
 import com.github.gfx.android.orma.Selector;
 import com.github.gfx.android.orma.annotation.Experimental;
+import com.github.gfx.android.orma.rx.RxRelation;
 
 import android.content.Context;
 import android.support.annotation.CheckResult;
@@ -46,7 +47,7 @@ public abstract class OrmaRecyclerViewAdapter<Model, VH extends RecyclerView.Vie
 
     protected final OrmaAdapter<Model> delegate;
 
-    public OrmaRecyclerViewAdapter(@NonNull Context context, @NonNull Relation<Model, ?> relation) {
+    public OrmaRecyclerViewAdapter(@NonNull Context context, @NonNull RxRelation<Model, ?> relation) {
         this(new OrmaAdapter<>(context, relation));
     }
