+++ processor/src/main/java/com/github/gfx/android/orma/processor/generator/RelationWriter.java
@@ -29,6 +29,7 @@ import com.squareup.javapoet.CodeBlock;
 import com.squareup.javapoet.FieldSpec;
 import com.squareup.javapoet.MethodSpec;
 import com.squareup.javapoet.ParameterSpec;
+import com.squareup.javapoet.ParameterizedTypeName;
 import com.squareup.javapoet.TypeSpec;
 
 import java.util.ArrayList;
@@ -71,7 +72,11 @@ public class RelationWriter extends BaseWriter {
             classBuilder.addAnnotation(Annotations.suppressWarnings("rawtypes"));
         }
         classBuilder.addModifiers(Modifier.PUBLIC);
-        classBuilder.superclass(Types.getRelation(schema.getModelClassName(), getTargetClassName()));
+
+        ParameterizedTypeName relationType = context.generationOption.isRxJavaSupport()
+                ? Types.getRxRelation(schema.getModelClassName(), getTargetClassName())
+                : Types.getRelation(schema.getModelClassName(), getTargetClassName());
+        classBuilder.superclass(relationType);
 
         classBuilder.addField(FieldSpec.builder(schema.getSchemaClassName(), "schema", Modifier.FINAL).build());
